+++ mp_reduce_2k_l.c
@@ -11,7 +11,7 @@ mp_err mp_reduce_2k_l(mp_int *a, const mp_int *n, const mp_int *d)
 {
    mp_int q;
    mp_err err;
-   int    p;
+   unsigned long p;
 
    if ((err = mp_init(&q)) != MP_OKAY) {
       return err;
